requirements.txt
click==6.6
cookies==2.2.1
Flask==0.11.1
geopy==1.11.0
giphypop==0.2
GroupyAPI==0.6.6
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.5.3
python-forecastio==1.3.4
requests==2.3.0
responses==0.5.1
six==1.10.0
Werkzeug==0.11.10
wheel==0.24.0
